Rename and split client models into separate files
