+++ internal/clients/instance_type.go
@@ -1,16 +1,8 @@
 package clients
 
 import (
-	"bytes"
-	"embed"
-	"errors"
-	"fmt"
-	"os"
-	"path/filepath"
 	"strconv"
 	"strings"
-
-	"gopkg.in/yaml.v3"
 )
 
 type InstanceTypeName string
@@ -104,224 +96,4 @@ func (it *InstanceType) SetMemoryFromBytes(memKib int64) {
 
 func (it *InstanceType) SetEphemeralStorageFromMB(storageMb int64) {
 	it.EphemeralStorageGB = storageMb / 1000
-}
-
-// RegisteredInstanceTypes holds all details about instance types.
-type RegisteredInstanceTypes struct {
-	types map[InstanceTypeName]*InstanceType `yaml:"registered_types"`
-}
-
-func NewRegisteredInstanceTypes() *RegisteredInstanceTypes {
-	return &RegisteredInstanceTypes{
-		types: make(map[InstanceTypeName]*InstanceType),
-	}
-}
-
-func (rit *RegisteredInstanceTypes) Register(it InstanceType) {
-	rit.types[it.Name] = &it
-}
-
-func (rit *RegisteredInstanceTypes) Get(name InstanceTypeName) *InstanceType {
-	return rit.types[name]
-}
-
-func (rit *RegisteredInstanceTypes) Load(buffer []byte) error {
-	err := yaml.Unmarshal(buffer, &rit.types)
-	if err != nil {
-		return fmt.Errorf("unable to unmarshal registered instance types: %w", err)
-	}
-
-	return nil
-}
-
-func (rit *RegisteredInstanceTypes) LoadSupported(buffer []byte) error {
-	supportedList := make([]InstanceTypeName, 0)
-	err := yaml.Unmarshal(buffer, &supportedList)
-	if err != nil {
-		return fmt.Errorf("unable to unmarshal supported instance types: %w", err)
-	}
-	supported := make(map[InstanceTypeName]bool, len(supportedList))
-	for _, name := range supportedList {
-		supported[name] = true
-	}
-
-	for itn, it := range rit.types {
-		if _, ok := supported[itn]; ok {
-			it.Supported = true
-		} else {
-			it.Supported = false
-		}
-	}
-
-	return nil
-}
-
-func (rit *RegisteredInstanceTypes) Save(filename string) error {
-	return compareAndMarshal(filename, rit.types)
-}
-
-func (rit *RegisteredInstanceTypes) Print(typeName string) {
-	if typeName != "" {
-		it, ok := rit.types[InstanceTypeName(typeName)]
-		if ok {
-			fmt.Println(it.String())
-		} else {
-			fmt.Println("Not found")
-		}
-	} else {
-		for _, v := range rit.types {
-			fmt.Println(v.String())
-		}
-		fmt.Printf("Total: %d\n", len(rit.types))
-	}
-}
-
-// RegionalTypeAvailability type is used to capture available instance types per
-// region and zone.
-type RegionalTypeAvailability struct {
-	types map[string][]InstanceTypeName
-}
-
-const regionSeparator = "_"
-
-func NewRegionalInstanceTypes() *RegionalTypeAvailability {
-	return &RegionalTypeAvailability{
-		types: make(map[string][]InstanceTypeName),
-	}
-}
-
-var UnknownRegionZoneCombinationErr error = errors.New("unknown region and zone combination")
-
-func (rit *RegionalTypeAvailability) NamesForZone(region, zone string) ([]InstanceTypeName, error) {
-	result, ok := rit.types[region+regionSeparator+zone]
-	if !ok {
-		return nil, UnknownRegionZoneCombinationErr
-	}
-	return result, nil
-}
-
-func (rit *RegionalTypeAvailability) Add(region, zone string, it InstanceType) {
-	raz := region + regionSeparator + zone
-	if _, ok := rit.types[raz]; !ok {
-		rit.types[raz] = make([]InstanceTypeName, 0)
-	}
-	rit.types[raz] = append(rit.types[raz], it.Name)
-}
-
-func (rit *RegionalTypeAvailability) Save(directory string) error {
-	for key, value := range rit.types {
-		filename := filepath.Join(directory, key+".yaml")
-		err := compareAndMarshal(filename, value)
-		if err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func (rit *RegionalTypeAvailability) Load(fsTypes embed.FS, path string) error {
-	rit.types = make(map[string][]InstanceTypeName)
-
-	dirEntries, err := fsTypes.ReadDir(path)
-	if err != nil {
-		return fmt.Errorf("unable to read availability dir: %w", err)
-	}
-
-	for _, dirEntry := range dirEntries {
-		if dirEntry.IsDir() {
-			continue
-		}
-		file := filepath.Join(path, dirEntry.Name())
-		buffer, err := fsTypes.ReadFile(file)
-		if err != nil {
-			return fmt.Errorf("unable to read availability file %s: %w", file, err)
-		}
-		key := strings.TrimSuffix(dirEntry.Name(), ".yaml")
-		var value []InstanceTypeName
-		err = yaml.Unmarshal(buffer, &value)
-		if err != nil {
-			return fmt.Errorf("unable to unmarshal availability file %s: %w", file, err)
-		}
-		rit.types[key] = value
-	}
-
-	return nil
-}
-
-var RegionAndZoneSplitErr = errors.New("unable to split region and zone for")
-
-func splitRegionZone(str string) (string, string, error) {
-	result := strings.Split(str, regionSeparator)
-	if len(result) != 2 {
-		return "", "", fmt.Errorf("%w: %s", RegionAndZoneSplitErr, str)
-	}
-	return result[0], result[1], nil
-}
-
-func (rit *RegionalTypeAvailability) Print(fRegion, fZone string) {
-	for raz, names := range rit.types {
-		region, zone, err := splitRegionZone(raz)
-		if err != nil {
-			panic(err)
-		}
-		if (fRegion == "" && fZone == "") ||
-			(fRegion == region && fZone == "") ||
-			(fRegion == region && fZone == zone) ||
-			(fRegion == "all" && fZone == "") {
-			fmt.Printf("Region '%s' availability zone '%s':\n", region, zone)
-			sb := strings.Builder{}
-			for _, name := range names {
-				sb.WriteString(name.String())
-				sb.WriteString(", ")
-			}
-			fmt.Println(sb.String())
-			fmt.Println("")
-		}
-	}
-}
-
-type InstanceTypeInfo struct {
-	RegisteredTypes      RegisteredInstanceTypes
-	RegionalAvailability RegionalTypeAvailability
-}
-
-func (iii *InstanceTypeInfo) InstanceTypesForZone(region, zone string, supported *bool) ([]*InstanceType, error) {
-	result := make([]*InstanceType, 0, 64)
-	names, err := iii.RegionalAvailability.NamesForZone(region, zone)
-	if err != nil {
-		return nil, err
-	}
-	for _, name := range names {
-		rt := iii.RegisteredTypes.Get(name)
-		if supported != nil && *supported != rt.Supported {
-			continue
-		}
-		result = append(result, rt)
-	}
-	return result, nil
-}
-
-func compareAndMarshal(filename string, obj any) error {
-	newBuffer, err := yaml.Marshal(obj)
-	if err != nil {
-		return fmt.Errorf("unable to marshal instance types: %w", err)
-	}
-
-	oldBuffer, err := os.ReadFile(filename)
-	if os.IsNotExist(err) {
-		oldBuffer = make([]byte, 0, 0)
-	} else if err != nil {
-		return fmt.Errorf("unable to read instance types: %w", err)
-	}
-
-	if !bytes.Equal(newBuffer, oldBuffer) {
-		/* #nosec */
-		err = os.WriteFile(filename, newBuffer, 0644)
-		if err != nil {
-			return fmt.Errorf("unable to save instance types: %w", err)
-		}
-	}
-
-	return nil
 }
